+++ packages/model-service/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.26.1](https://github.com/crossmodel/crossmodel-core/compare/v0.26.0...v0.26.1) (2026-02-06)
+
+**Note:** Version bump only for package @crossmodel/model-service
+
+
+
+
+
+# [0.26.0](https://github.com/crossmodel/crossmodel-core/compare/v0.25.1...v0.26.0) (2026-02-05)
+
+**Note:** Version bump only for package @crossmodel/model-service
+
+
+
+
+
+## [0.25.1](https://github.com/crossmodel/crossmodel-core/compare/v0.25.0...v0.25.1) (2026-02-04)
+
+**Note:** Version bump only for package @crossmodel/model-service
+
+
+
+
+
 # [0.25.0](https://github.com/crossmodel/crossmodel-core/compare/v0.24.1...v0.25.0) (2026-02-03)
 
 **Note:** Version bump only for package @crossmodel/model-service